Update profile page UI
