+++ frontend/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "framer-motion": "^12.23.12",
+    "lucide-react": "^0.542.0",
     "next": "15.5.2",
     "next-auth": "^5.0.0-beta.29",
     "react": "19.1.0",
