+++ frontend/src/app/profile/page.tsx
@@ -3,6 +3,8 @@
 
 import { useEffect, useState } from 'react';
 import { getUserProfile } from '../lib/profile-api';
+import { Pencil, User, Tag, Scroll, IdCard, Bell, Settings, Bot, LogOut, ChevronLeft } from 'lucide-react';
+import Link from 'next/link';
 
 export default function ProfilePage() {
   const [user, setUser] = useState<any>(null);
@@ -30,39 +32,57 @@ export default function ProfilePage() {
 
   if (!user) return <p className="p-4">Could not load profile</p>;
 
-  return (
-    <div className="max-w-md mx-auto mt-10 p-6 border rounded-lg shadow">
-      <h1 className="text-2xl font-bold mb-4">My Profile</h1>
-      <div className="space-y-2">
-        <p><strong>ID:</strong> {user._id}</p>
-        <p><strong>Name:</strong> {user.name}</p>
-        <p><strong>Email:</strong> {user.email}</p>
-        <p><strong>Role:</strong> 
-        <span className={`ml-2 px-2 py-1 rounded text-sm font-medium ${
-          user.role === 'seller' 
-            ? 'bg-blue-100 text-blue-800' 
-            : 'bg-green-100 text-green-800'
-        }`}>
-          {user.role}
-        </span>
+  const menuItems = [
+    { icon: <Scroll size={20} />, label: 'Privacy Policy', href: `/privacy-policy`},
+    { icon: <IdCard size={20} />, label: 'Payment Methods', href: `/payment-methods` },
+    { icon: <Bell size={20} />, label: 'Notification', href: `/notifications` },
+    { icon: <Settings size={20} />, label: 'Settings', href: `/settings` },
+    { icon: <Bot size={20} />, label: 'Help', href: `/help`},
+    { icon: <LogOut size={20} />, label: 'Logout', href: `/logout` },
+  ];
 
-        </p>
+  return (
+    <div className="bg-white dark:bg-gray-800 text-black dark:text-white min-h-screen">
+      <div className="flex justify-between items-center p-4">
+        <ChevronLeft size={24} className="text-black dark:text-white" />
+        <h1 className="text-xl font-bold text-pink-600">My Profile</h1>
+        <Link href={`/profile/update`}>
+          <Pencil size={24} className="text-pink-600" />
+        </Link>
       </div>
-      
-      {/* Role-specific content */}
-      {user.role === 'seller' && (
-        <div className="mt-4 p-4 bg-blue-50 rounded">
-          <h3 className="font-semibold text-blue-800">Seller Dashboard</h3>
-          <p className="text-sm text-blue-600">Manage your products and orders</p>
+
+      <div className="flex flex-col items-center mt-8">
+        <div className="relative w-32 h-32 rounded-full overflow-hidden border-2 border-[#F3E0DB]">
+          <User className="w-full h-full object-cover" />
         </div>
-      )}
-      
-      {user.role === 'buyer' && (
-        <div className="mt-4 p-4 bg-green-50 rounded">
-          <h3 className="font-semibold text-green-800">Buyer Profile</h3>
-          <p className="text-sm text-green-600">Browse and purchase products</p>
+        <h2 className="text-2xl font-bold mt-4 text-black dark:text-white">Name: {user.name}</h2>
+        <p className="text-2xl text-black dark:text-white mt-1">Email: {user.email}</p>
+        <p className="text-2xl text-black dark:text-white mt-1">Role: {user.role}</p>
+      </div>
+
+      <div className="flex justify-around p-4 mt-8 bg-[#3F2E31] rounded-lg mx-6 shadow-lg">
+        <div className="flex flex-col items-center">
+          <User size={24} className="text-white" />
+          <Link href={`/profile`} className="text-sm mt-2 text-white">Profile</Link>
         </div>
-      )}
+        <div className="flex flex-col items-center">
+          <Tag size={24} className="text-white" />
+          <Link href={`/wishlist/${user._id}`} className="text-sm mt-2 text-white">My wishlist</Link>
+        </div>
+        <div className="flex flex-col items-center">
+          <IdCard size={24} className="text-white" />
+          <Link href={`/orders/${user._id}`} className="text-sm mt-2 text-white">My Orders</Link>
+        </div>
+      </div>
+
+      <div className="mt-8 mx-6 space-y-4">
+        {menuItems.map((item, index) => (
+          <div key={index} className="flex items-center space-x-4">
+            <div className="p-3 bg-[#3F2E31] rounded-full text-[#F3E0DB]">{item.icon}</div>
+            <Link href={item.href} className="text-base text-[#F3E0DB]">{item.label}</Link>
+          </div>
+        ))}
+      </div>
     </div>
   );
 }
